requirements.txt
pip==21.3.1
pandas==1.2.3
panel
langchain==0.0.18
openai
# geogif
fire
psycopg2-binary
rasterio
# stackstac
odc-stac
botocore
odc-algo
psycopg2
rasterio
rioxarray
spyndex
xarray_leaflet